pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "advisor-scheduler"
version = "0.1.0"
description = "Advisor appointment scheduler prototype (Phase 1)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "httpx>=0.27",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
]

[project.optional-dependencies]
llm = ["google-genai>=0.8.0"]
mcp = [
  "fastmcp>=2.0",
  "google-api-python-client>=2.120",
  "google-auth>=2.29",
  "google-auth-oauthlib>=1.2",
  "google-auth-httplib2>=0.2",
]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "advisor-scheduler[llm]",
  "advisor-scheduler[mcp]",
]

[project.scripts]
advisor-google-mcp = "advisor_scheduler.integrations.google_workspace.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
  "mcp: live Google MCP integration (requires credentials and MCP server)",
]
